Revert "Add obsolete lists"
This reverts commit 1bd0f3795c5d603e318fd2a0af600badf59a4960.

This was created to migrate Turris OS 4.0-alpha* releases. The first
supported version 4.0-beta1 already contains correct path with board
specific path so there is no need to preserve these obsolete ones.
Non the less they are also wrong with additional introduced changes.
They point to pkglists in invalid location.

This requires also removal of code on build server to not use --obsolete
option and to not deploy those.